Added replace with background sampler (#72)
Summary:
Pull Request resolved: https://github.com/facebookresearch/AugLy/pull/72

Added new augmentation replace_with_background for videos, which replaces the beginning and end of the source video with the background video such that `source_percentage` of the final video is taken by the source video and the final output video is the same duration of the source video.

Also added this new transform to the sampling pipeline for global vce embedding, where it, insert_in_background, and a random spatial transform are all sampled with equal probability.

Differential Revision: D29050824

fbshipit-source-id: 223888114e0422bc1ad4340e64f8f3e5958656e6