Add blurry_mask augmentation (#104)
Summary:
Pull Request resolved: https://github.com/facebookresearch/AugLy/pull/104

Another augmentation we used for ISC that might be useful for others.

Note: there are a lot of constants hardcoded in the `make_mask()` function. Since Matthijs wrote that, I'll wait until he's back from PTO to ask his opinion on if we should expose any of them. I tried modifying them in a notebook and the effect on the augmented image wasn't clear to me.

Differential Revision: D30218994

fbshipit-source-id: 2a15dc06f77c3d4626dc51bd111ddc89640342d6