Add resize_src_to_match_template option to screenshot augmentation
Summary: In the ISC dataset we wanted this ability because some of the images were a lot bigger than others, so the screenshot templates would blow up. Setting `resize_src_to_match_template = True` means that the input image will be resized to fit in the template image's bounding box, meaning that the augmented image will have the same dimensions as the raw template image.

Reviewed By: jbitton

Differential Revision: D30221154

fbshipit-source-id: a251e85e868eb30a270de1e5cf8e0c0b22ff3e54