Add error messages for easier debugging (#151)
Summary:
Pull Request resolved: https://github.com/facebookresearch/AugLy/pull/151

The `False is not True` error that occurs when the expected media/metadata doesn't match is hard to parse for our users. Let's add more detailed messaging such that it's easier to debug augmentations!

This change doesn't apply to the text library since those tests don't have a unified base test etc :)

Reviewed By: zpapakipos

Differential Revision: D32604698

fbshipit-source-id: 5f6e8175903046b94c7b87a1e1eeb849eaa3b580