Final Sphinx documentation w/ ReadTheDocs (#172)
Summary:
- [x] I have read CONTRIBUTING.md to understand how to contribute to this repository :)

<Please summarize what you are trying to achieve, what changes you made, and how they acheive the desired result.>

### What This Is
This is comprehensive documentation for AugLy with Sphinx docstring formatting under the hood with a Read the Docs theme to make augmentation parameters, return types, etc more readable and user-friendly.

### How It Works
Sphinx is a documentation generator commonly used by the Python community. It also has its own docstring format.

Internal AugLy docstrings utilize tags such as `param` or `returns` for labeling due to internal Facebook convention. However, Sphinx does not recognize these tags, opting for `:` in favor of `@`, among other changes.

Luckily for us, Python docstring format `epytext` is **very** similar to AugLy (credit Cloud9c), meaning that we can claim that we use epytext formatting and then convert it to `sphinx` when necessary.

Another problem: Sphinx requires explicit types labeled in the form of `:type` and `:rtype` to display types once documentation is rendered. However, Python typehints (which AugLy uses generously) are not natively supported. Therefore we use an extension to Sphinx that autodetects typehints and adds them on the fly.

In the end, Sphinx uses the module structure specified in `docs/source` with rST (filetype similar to Markdown) to generate a table of contents for our final documentation.

### How to Build Documentation Locally

1. Clone repository. `git clone github.com/facebookresearch/AugLy`
2. Install all requirements for both the library and documentation-specific dependencies `cd docs && pip install -r requirements.txt`
3. Make sure you are in the `docs` subdirectory. Then run `make html` to generate documentation. If you want to delete all these later, you can run `make clean`.
4. Navigate to `docs/build` and open `index.html`

### Generating new documentation later

1. Sphinx can detect new files added to the `augly` subdirectory and add their `.rst` files accordingly, but the process of detection needs to be triggered manually. Run `sphinx-apidoc -o ./source ../augly` in the root directory to do so, and update the `toctree` in `index.rst` if necessary.
2. Edited a docstring and want to see the same changes be reflected in the published documentation? No worries, this will be automatically completed, an overview of which is provided below.

### Integration with ReadTheDocs
1. This documentation uses Sphinx's ReadTheDocs theme to make it easy to publish documentation on RTD's site.
2. Create a Github webhook to detect pushes made to the repository so documentation can rebuild.
3. The `.readthedocs.yml` file specifies the configuration for these builds. By default `ffmpeg` and `libsndfile1` are based on C and are required as prerequisites before requirements in `docs/requirements.txt` are installed (as RTD uses Ubuntu behind the scenes).
4. Docstrings aren't stored in the `docs` subdirectory at all, so all are read from the source folder. Updating the docstrings in `augly/<modality>` will be sufficient.

Pull Request resolved: https://github.com/facebookresearch/AugLy/pull/172

Reviewed By: evtimovi

Differential Revision: D33028278

Pulled By: jbitton

fbshipit-source-id: 6cafa5c5ced17d40aa7ce4bbc028e3c69efa4f71