Move test data (#203)
Summary:
## Related Issue
Fixes https://github.com/facebookresearch/AugLy/issues/196

- [x] I have read CONTRIBUTING.md to understand how to contribute to this repository :)

Moving the test data out of the `assets` dir which is all packaged with the pypi package, and into the `tests` dir which is excluded from packaging. This will decrease the size of the pypi package in the future.

## Unit Tests
See `test_python` workflow.

Pull Request resolved: https://github.com/facebookresearch/AugLy/pull/203

Reviewed By: jbitton

Differential Revision: D34814217

Pulled By: zpapakipos

fbshipit-source-id: 5a88935391cb878e9259a74c1659ceccd07e7368