Fixed bug in segments calculation for replace_in_background
Summary:
Changed the logic of updating segments for `replace_in_background` to account for time cropping. When applying this transform in isolation, the `src_segments` and `dst_segments` should be equal, which was not the case.

Additional change: adjusted the way `new_start` incorporates potential speed-ups in the code to improve consistency.

Reviewed By: zpapakipos

Differential Revision: D38070720

fbshipit-source-id: 7c164900703cca7ee4e7681899618b53422b35ab