Resample background noise to seed audio's sample rate in add_background_noise (#244)
Summary:
Pull Request resolved: https://github.com/facebookresearch/AugLy/pull/244

`add_background_noise` allows the caller to specify a path to load a noise waveform from. It's possible that the noise waveform and the seed audio to augment don't have the same sample rate. When this is the case, `add_background_noise` silently and unexpectedly adds the noise and audio with mismatched sample rates. To address this, this diff modifies `add_background_noise` to resample the noise waveform to the seed audio's sample rate.

Reviewed By: xiaohui-zhang

Differential Revision: D45331594

fbshipit-source-id: 5528244ceca9ade86d65817c02c292a0c011b9bb