Add ranking numbers transformation
Summary: This diff adds a new augmentation function to the AugLy library called `ranking_numbers`. This function takes an image and a dictionary of rankings and positions, and adds text to the image in the form of ranking numbers. The function can also take additional parameters such as font, font size, color, stroke width, and stroke fill. The function is added to the `__init__.py` file and the `functional.py` file, and is also added to the list of available augmentations in the `__init__.py` file. The function is tested in the `functional_unit_test.py` file

Reviewed By: Zhourj

Differential Revision: D79221943

fbshipit-source-id: 0c38875aff974fb5ec348436402ea3704de7c260